pyproject.toml
[project]
name = "django_starter"
version = "0.1.0"
description = "A starter template for Django projects"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "dj-rest-auth>=6.0.0",
    "django>=5.2.4",
    "django-allauth>=65.10.0",
    "django-celery-beat>=2.8.1",
    "django-ckeditor-5>=0.2.18",
    "django-cors-headers>=4.7.0",
    "django-environ==0.11.2",
    "django-filter>=25.1",
    "django-json-widget>=2.0.3",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "drf-yasg>=1.21.10",
    "fabric>=3.2.2",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
    "sentry-sdk>=2.34.1",
    "shortuuid>=1.0.13",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "coverage>=7.10.2",
    "django-debug-toolbar>=6.0.0",
    "django-extensions>=4.1",
    "watchdog>=6.0.0",
    "werkzeug>=3.1.3",
]
